feat: Add breast cancer prediction form
This commit adds a new HTML form for breast cancer prediction. The form is designed with input fields for various features related to breast cancer. When the form is submitted, the input values are used to make a prediction using a pre-trained machine learning model. The prediction result is displayed on a new result page.